Правки тального і линовиці на заводі
git-svn-id: file:///tmp/qt4/Linovitsa@338 76bea514-bf2f-4ad3-ac3e-82dc02406d2c

+++ kf-1000/ioserv/reportwr.h
@@ -20,7 +20,7 @@ private:
     QVector<IoDev*> &src;
 
     QVector<int> lastState;
-    QVector<QVector<int> > data;
+    QVector<QVector<double> > data;
 
 };
